* Splitting the SDK simulator into separate file so the interface is consistent. * Separate exceptions raised for different issues with Camera creation.
