+++ pocs/camera/simulator/__init__.py
@@ -1,0 +1,1 @@
+from pocs.camera.simulator.dslr import Camera
