+++ pocs/camera/simulator/dslr.py
@@ -9,7 +9,6 @@ from astropy import units as u
 from astropy.io import fits
 
 from pocs.camera import AbstractCamera
-from pocs.camera.sdk import AbstractSDKDriver, AbstractSDKCamera
 from pocs.utils.images import fits as fits_utils
 from pocs.utils import get_quantity_value
 
@@ -91,26 +90,3 @@ class Camera(AbstractCamera):
 
         self.logger.debug("Headers updated for simulated image.")
         return file_path
-
-
-class SDKDriver(AbstractSDKDriver):
-    def __init__(self, library_path=None, **kwargs):
-        # Get library loader to load libc, which should usually be present...
-        super().__init__(name='c', library_path=library_path, **kwargs)
-
-    def get_SDK_version(self):
-        return "Simulated SDK Driver v0.001"
-
-    def get_cameras(self):
-        cameras = {'SSC007': 'DEV_USB0',
-                   'SSC101': 'DEV_USB1',
-                   'SSC999': 'DEV_USB2'}
-        return cameras
-
-
-class SDKCamera(AbstractSDKCamera, Camera):
-    def __init__(self,
-                 name='Simulated SDK camera',
-                 driver=SDKDriver,
-                 *args, **kwargs):
-        super().__init__(name, driver, *args, **kwargs)
